Extract enums into separate headers (#764)
