+++ src/Manifest.cpp
@@ -28,8 +28,8 @@
 #include "File.hpp"
 #include "Hash.hpp"
 #include "Logging.hpp"
+#include "Sloppiness.hpp"
 #include "StdMakeUnique.hpp"
-#include "ccache.hpp"
 #include "fmtmacros.hpp"
 #include "hashutil.hpp"
