+++ src/argprocessing.hpp
@@ -19,7 +19,7 @@
 #pragma once
 
 #include "Args.hpp"
-#include "Statistics.hpp"
+#include "Statistic.hpp"
 
 #include "third_party/nonstd/optional.hpp"
