+++ src/hashutil.cpp
@@ -23,8 +23,8 @@
 #include "Context.hpp"
 #include "Hash.hpp"
 #include "Logging.hpp"
+#include "Sloppiness.hpp"
 #include "Stat.hpp"
-#include "ccache.hpp"
 #include "execute.hpp"
 #include "fmtmacros.hpp"
 #include "macroskip.hpp"
