+++ unittest/test_Statistics.cpp
@@ -16,6 +16,7 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
+#include "../src/Statistic.hpp"
 #include "../src/Statistics.hpp"
 #include "../src/Util.hpp"
 #include "../src/fmtmacros.hpp"
